Factor out threadhandleobject.c.
